// src/app/core/models/categories/transformed/category.data.ts
export type Categories =
  | 'Area Control'
  | 'Ameritrash'
  | 'Campaign'
  | 'Deckbuilder'
  | 'Deckbuilding'
  | 'Deck Construction'
  | 'Dexterity'
  | 'Drafting'
  | 'Dungeon Crawler'
  | 'Engine Builder'
  | 'Eurogame'
  | 'Party Game'
  | 'Push your Luck'
  | 'Roll and Move'
  | 'Roll and Write'
  | 'Social Deduction'
  | 'Storytelling'
  | 'Worker-placement'
  | 'Wargame';